Move SessionCollection to plex module
